+++ compiled/lib/sfdx-client.d.ts
@@ -6,6 +6,23 @@ export declare enum ApiKind {
     COMPOSITE = "composite"
 }
 export declare class SfdxClient {
+    static readonly metaDataInfo: {
+        ContentVersion: {
+            MetaName: string;
+            DataName: string;
+            Filename: string;
+        };
+        Document: {
+            MetaName: string;
+            DataName: string;
+            Filename: string;
+        };
+        Attachment: {
+            MetaName: string;
+            DataName: string;
+            Filename: string;
+        };
+    };
     private static defailtIdField;
     private headers;
     private orgAliasOrUsername;
@@ -22,6 +39,7 @@ export declare class SfdxClient {
     updateByRecord(metaDataType: string, record: any, recordIdField?: string, apiKind?: ApiKind): Promise<RestResult>;
     updateByRecords(metaDataType: string, records: any[], recordIdField?: string, apiKind?: ApiKind): AsyncGenerator<RestResult, void, void>;
     doMultiPart(action: RestAction, record: any, filePath: string, validStatusCodes?: number[]): Promise<RestResult>;
+    postObjectMultipart(objectName: string, objectRecord: any, fileName: string, filePath: string): Promise<any>;
     do(action: RestAction, metaDataType: string, records?: any[], recordIdField?: string, apiKind?: ApiKind, validStatusCodes?: number[]): AsyncGenerator<RestResult, void, void>;
     doComposite(action: RestAction, record: any, validStatusCodes?: number[]): Promise<RestResult>;
     getMaxApiVersion(): Promise<string>;
