+++ compiled/lib/sfdx-core.d.ts
@@ -1,3 +1,4 @@
+/// <reference types="node" />
 export declare class SfdxCore {
     static ASTERIX: string;
     static MAIN: string;
