+++ messages/acu-pack.json
@@ -120,7 +120,8 @@
         "metadataFlagDescription": "The MetaData Type name to upload",
         "commandDescription": "\r\nUploads ContentVersion files using a multi-part message when necessary.",
         "recordsFlagDescription": "The Path to the file (CSV) containing the ContentVersion record data to upload",
-        "columnsFlagDescription": "A comma seperated list of the columns to use from the CSV file. If not specififed, all the columns are used."
+        "columnsFlagDescription": "A comma seperated list of the columns to use from the CSV file. If not specififed, all the columns are used.",
+        "allOrNothingFlagDescription": "Set this flag to stop the upload process on the first error"
       }
     },
     "get": {
